python/flask_demo/requirements.txt
# Requirements automatically generated by pigar.
# https://github.com/Damnever/pigar

# app/common/base.py: 13,14,15
# app/common/decorator.py: 8,9,10,11,12
# app/config.py: 10
# app/models.py: 16
# app/views/api.py: 10
# app/views/html.py: 12,13,14,15
# app/views/test.py: 13,14,15,16,17
Flask == 0.12

# app/config.py: 11
Flask_Cors == 3.0.2

# app/config.py: 13
# app/views/html.py: 16,17,18
Flask_Login == 0.4.1

# app/run.py: 14
Flask_Restless == 0.17.0

# app/config.py: 12
Flask_SQLAlchemy == 2.1

# app/common/md.py: 8
Markdown == 2.6.9

# app/common/base.py: 22
PyMySQL == 0.7.9

# app/common/base.py: 17,18
# app/models.py: 13,14,15
SQLAlchemy == 1.1.5

# app/config.py: 14
Werkzeug == 0.12.1

# app/common/elasticsearch.py: 8,9,10,11
elasticsearch == 5.2.0

# app/common/md.py: 9
markdown2 == 2.3.4

# app/common/md.py: 7
mistune == 0.7.4

# app/common/aliyun.py: 7
# app/common/qiniu.py: 7
oss2 == 2.3.3

# app/common/security.py: 11
pycrypto == 2.6.1

# app/common/aliyun.py: 8
# app/common/email_utils.py: 12
# app/common/qiniu.py: 8
# app/common/third/juhe.py: 7
# app/common/third/wx.py: 14
requests == 2.13.0

# app/common/third/wx.py: 15
xmltodict == 0.11.0
